requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
motor==2.5.1
pymongo<4.0
Pillow==9.5.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo[srv]
dnspython
dotenv==0.0.5
